# backend/openapi/TidySense.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TidySense | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/goals": {
      "get": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfGoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfGoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfGoalDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{id}": {
      "get": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{id}/review": {
      "post": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewGoalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewGoalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewGoalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{id}/terminal-preview": {
      "post": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{id}/terminal": {
      "post": {
        "tags": [
          "Goals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiProblemDto": {
        "required": [
          "type",
          "title",
          "status",
          "code",
          "messageKey",
          "retryable",
          "traceId",
          "correlationId",
          "detail",
          "errors",
          "entityId",
          "expectedVersion",
          "currentVersion"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "traceId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "entityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currentVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "CreateGoalRequest": {
        "required": [
          "title",
          "desiredOutcome",
          "targetDate",
          "reviewDate"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "desiredOutcome": {
            "type": "string"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
...
      "CursorPageDtoOfGoalDto": {
        "required": [
          "items",
          "page"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoalDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoDto"
          }
        }
      },
...
      "GoalDto": {
        "required": [
          "id",
          "title",
          "desiredOutcome",
          "status",
          "targetDate",
          "reviewDate",
          "reviewDateSource",
          "lastContinuationDecisionAt",
          "source",
          "version",
          "createdAt",
          "updatedAt",
          "terminalAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "desiredOutcome": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": "string",
            "format": "date"
          },
          "reviewDateSource": {
            "type": "string"
          },
          "lastContinuationDecisionAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "terminalAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PageInfoDto": {
        "required": [
          "nextCursor",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "ReviewGoalRequest": {
        "required": [
          "decision",
          "reviewDate",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "decision": {
            "type": "string"
          },
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "TerminalBlockerDto": {
        "required": [
          "resourceType",
          "resourceId",
          "status",
          "version"
        ],
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "TerminalCascadeDto": {
        "required": [
          "resourceType",
          "resourceId",
          "resultingStatus",
          "version"
        ],
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "resultingStatus": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "TerminalCommandRequest": {
        "required": [
          "targetStatus",
          "expectedVersion",
          "previewHash"
        ],
        "type": "object",
        "properties": {
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "previewHash": {
            "type": "string"
          }
        }
      },
      "TerminalPreviewDto": {
        "required": [
          "entityId",
          "entityType",
          "currentStatus",
          "targetStatus",
          "expectedVersion",
          "canApply",
          "blockers",
          "cascades",
          "previewHash"
        ],
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string"
          },
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "canApply": {
            "type": "boolean"
          },
          "blockers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TerminalBlockerDto"
            }
          },
          "cascades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TerminalCascadeDto"
            }
          },
          "previewHash": {
            "type": "string"
          }
        }
      },
      "TerminalPreviewRequest": {
        "required": [
          "targetStatus",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "UpdateGoalRequest": {
        "required": [
          "title",
          "desiredOutcome",
          "targetDate",
          "reviewDate",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "desiredOutcome": {
            "type": "string"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}